+++ src/main/kotlin/dev/mayaqq/estrogen/client/content/screen/EstrogenMenuScreen.kt
@@ -4,13 +4,11 @@ import dev.mayaqq.cynosure.client.utils.pushPop
 import dev.mayaqq.cynosure.client.utils.translate
 import dev.mayaqq.cynosure.helpers.McClient
 import dev.mayaqq.cynosure.helpers.openUri
-import dev.mayaqq.cynosure.text.CommonText
 import dev.mayaqq.cynosure.text.Text
 import dev.mayaqq.cynosure.text.TextBuilder.append
 import dev.mayaqq.cynosure.text.TextStyle.bold
 import dev.mayaqq.cynosure.text.TextStyle.color
 import dev.mayaqq.cynosure.utils.colors.Color
-import dev.mayaqq.cynosure.utils.colors.Gray
 import dev.mayaqq.cynosure.utils.colors.LightBlue
 import dev.mayaqq.cynosure.utils.colors.Red
 import dev.mayaqq.cynosure.utils.colors.White
@@ -21,8 +19,7 @@ import dev.mayaqq.estrogen.id
 import net.minecraft.client.gui.GuiGraphics
 import net.minecraft.client.gui.components.Tooltip
 import net.minecraft.client.gui.screens.Screen
-import net.minecraft.client.resources.sounds.SimpleSoundInstance
-import net.minecraft.sounds.SoundEvents
+import kotlin.io.path.exists
 
 class EstrogenMenuScreen(previous: Screen?) : BaseEstrogenScreen(previous, Text.translatable("estrogen.screen.menu.title")) {
 
@@ -43,8 +40,7 @@ class EstrogenMenuScreen(previous: Screen?) : BaseEstrogenScreen(previous, Text.
         //TODO: open cosmetics screen
     }
     val bMemorial = EstrogenButton.Builder(EstrogenButton.TextRenderer(Text.translatable("estrogen.button.memorial"))) {
-        //TODO: Open memorial Screen
-        McClient.soundManager.play(SimpleSoundInstance.forUI(SoundEvents.CAT_AMBIENT, 1.0F))
+        McClient.setScreen(MemorialScreen(this))
     }
     val bClose = EstrogenButton.Builder(EstrogenButton.TextRenderer(Text.translatable("estrogen.button.close"))) {
         this.onClose()
@@ -70,17 +66,12 @@ class EstrogenMenuScreen(previous: Screen?) : BaseEstrogenScreen(previous, Text.
         val y = this.height / 4 + 48
         val rowHeight = 24
 
-        val memorialEnabled = GlobalStorage.getData(MOD_ID).resolve("/memorial/unlocked").toFile().exists()
+        val memorialEnabled = GlobalStorage.getData(MOD_ID).resolve("memorial").exists()
         if (!memorialEnabled) {
             bMemorial.disabled(true)
             bMemorial.tooltip(Tooltip.create(Text.of {
                 append(Text.translatable("estrogen.button.memorial.desc")) {
                     color = Red
-                }
-                append(CommonText.NEWLINE)
-                append(CommonText.NEWLINE)
-                append(Text.translatable("estrogen.generic.coming_soon")) {
-                    color = Gray
                 }
             }))
         }